refactor: move brandColors to dependency-free module
Addresses Copilot review on #550: brand-colors.ts now holds the values
with zero imports, so raw-hex consumers (WebView pages, Stripe
appearance, React Navigation themes) don't pull tamagui/createTamagui
into their module graph. tamagui.config.ts imports from it and also
re-exports it so the web webpack singleton alias keeps exposing
brandColors regardless of merge order with #548. Stale App.tsx JSDoc
updated.
