+++ packages/config/src/index.ts
@@ -1,4 +1,4 @@
 export { config } from "./tamagui.config";
 export type { AppConfig } from "./tamagui.config";
-export { brandColors } from "./tamagui.config";
-export type { BrandColor } from "./tamagui.config";
+export { brandColors } from "./brand-colors";
+export type { BrandColor } from "./brand-colors";
